# src/api/dto/base.go
package dto

type CreateUpdateCountryRequest struct {
	Name string `json:"name" binding:"required,alpha,min=3,max=20"`
}

// type CountryResponse struct {
// 	Id        int               `json:"id"`
// 	Name      string            `json:"name"`
// 	Cities    []CityResponse    `json:"cities,omitempty"`
// 	Companies []CompanyResponse `json:"companies,omitempty"`
// }

type CreateCityRequest struct {
	Name      string `json:"name" binding:"required,alpha,min=3,max=20"`
	CountryId int    `json:"countryId" binding:"required"`
}

type UpdateCityRequest struct {
	Name      string `json:"name,omitempty" binding:"alpha,min=3,max=20"`
	CountryId int    `json:"countryId,omitempty"`
}

// type CityResponse struct {
// 	Id      int             `json:"id"`
// 	Name    string          `json:"name"`
// 	Country CountryResponse `json:"country,omitempty"`
// }

// type CompanyResponse struct {
// 	Id      int             `json:"id"`
// 	Name    string          `json:"name"`
// 	Country CountryResponse `json:"country,omitempty"`
// }

type CountryResponse struct {
	Id   int    `json:"id"`
	Name string `json:"name"`
}